+++ src/components/error/NotFound.js
@@ -1,0 +1,12 @@
+import React, {Component} from 'react';
+
+export default class NotFound extends Component {
+    render() {
+        return (
+            <div>
+                <h2>404</h2>
+                <p>Not Found</p>
+            </div>
+        )
+    }
+}
