+++ src/css/global.js
@@ -180,7 +180,7 @@ opacity: 1;
   padding: 1em .5em;
 }
 
-div.movies {
+.movies {
   margin: 5em;
   display: flex;
   flex-flow: row wrap;
@@ -188,12 +188,12 @@ div.movies {
 }
 
 @media (max-width: 991px) {
-  div.movies {
+  .movies {
   margin: 2em;
   }
 }
 
-div.skeleton-movies {
+.skeleton-movies {
   display: flex;
   flex-flow: row wrap;
   justify-content: center;
@@ -241,7 +241,7 @@ div.skeleton-movies {
     width: 20%;
   }
 
-  div.movie {
+  .movie {
       width: 100%;
   }
 }
@@ -253,7 +253,7 @@ div.skeleton-movies {
     }
 }
 
-div.movie {
+.movie {
     width: 220px;
     min-height: 30vh;
     height: 100%;
@@ -263,7 +263,7 @@ div.movie {
     display: block;
   }
   
-  div.movie-image {
+  .movie-image {
     height: 80%;
     width: 100%;
     margin: auto;
@@ -272,7 +272,7 @@ div.movie {
     position: relative;
   }
   
-  div.movie-image img {
+  .movie-image img {
     box-shadow: 2px 7px 8px 2px ${({ theme }) => theme.movieBoxShadow};
     height: 20em;
     width: 100%;
@@ -284,7 +284,7 @@ div.movie {
     object-fit: cover;
   }
   
-  div.movie-image img:hover {
+  .movie-image img:hover {
     box-shadow: 2px -2px 8px 2px ${({ theme }) => theme.movieBoxShadow};
     transform: translateY(-3%);
   }
@@ -374,7 +374,7 @@ div.movie {
   /* Modal boostrap style End */
   /* skeleton loader styles */
   
-  div.skeleton-movie {
+  .skeleton-movie {
     width: 220px;
     min-height: 30vh;
     background: transparent;
@@ -383,7 +383,7 @@ div.movie {
     display: block;
   }
 
-  div.skeleton-movie-image {
+  .skeleton-movie-image {
     height: 310px;
     width: 100%;
     margin: auto;
@@ -416,7 +416,7 @@ div.movie {
     box-shadow: 0px 0px 7px -2px #121315;
   }
   
-  div.skeleton-movie__about p {
+  .skeleton-movie__about p {
     border-radius: 5px;
   }
   
@@ -561,7 +561,7 @@ select {
   color: #dcc80d !important;
 } /* Movie ard for small screens */
   @media (max-width: 561px) {
-  div.movie {
+  .movie {
     display: flex;
     width: 100%;
     height: initial;
@@ -570,12 +570,12 @@ select {
   }
 
 
-  div.movie-image {
+  .movie-image {
     height: 100%;
   }
 
 
-  div.movie-image img {
+  .movie-image img {
     height: 100%;
     border-bottom-left-radius: 0.3em;
     border-top-right-radius: initial;
@@ -608,7 +608,7 @@ select {
 
 
 
-  div.skeleton-movies {
+  .skeleton-movies {
     display: flex;
     flex-flow: column;
     flex-direction: column;
@@ -617,7 +617,7 @@ select {
     overflow: hidden;
   }
 
-  div.skeleton-movie {
+  .skeleton-movie {
     display: flex;
     width: 100%;
     flex-direction: row;
@@ -628,13 +628,13 @@ select {
     justify-content: space-between;
   }
 
-  div.skeleton-movie-image {
+  .skeleton-movie-image {
     height: 15em;
     width: 100%;
     margin: 0 .5em 0 0;
   }
 
-  div.sabout {
+  .sabout {
     width: 100%;
   }
 
@@ -645,7 +645,7 @@ select {
       right: -9em;
   }
 
-  div.skeleton-movie__about p {
+  .skeleton-movie__about p {
     border-radius: 5px;
   }
